+++ sdk/appplatform/azure-mgmt-appplatform/generated_samples/configuration_services_delete.py
@@ -29,14 +29,13 @@ def main():
         subscription_id="00000000-0000-0000-0000-000000000000",
     )
 
-    response = client.configuration_services.begin_delete(
+    client.configuration_services.begin_delete(
         resource_group_name="myResourceGroup",
         service_name="myservice",
         configuration_service_name="default",
     ).result()
-    print(response)
 
 
-# x-ms-original-file: specification/appplatform/resource-manager/Microsoft.AppPlatform/stable/2022-12-01/examples/ConfigurationServices_Delete.json
+# x-ms-original-file: specification/appplatform/resource-manager/Microsoft.AppPlatform/stable/2023-12-01/examples/ConfigurationServices_Delete.json
 if __name__ == "__main__":
     main()