+++ sdk/containerservice/azure-mgmt-containerservice/azure/mgmt/containerservice/_container_service_client.py
@@ -53,7 +53,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
     """
 
-    DEFAULT_API_VERSION = '2023-10-01'
+    DEFAULT_API_VERSION = '2023-11-01'
     _PROFILE_TAG = "azure.mgmt.containerservice.ContainerServiceClient"
     LATEST_PROFILE = ProfileDefinition({
         _PROFILE_TAG: {
@@ -61,7 +61,10 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             'container_services': '2019-04-01',
             'fleet_members': '2022-09-02-preview',
             'fleets': '2022-09-02-preview',
+            'machines': '2023-10-02-preview',
+            'managed_cluster_snapshots': '2023-10-02-preview',
             'open_shift_managed_clusters': '2019-04-30',
+            'operation_status_result': '2023-10-02-preview',
         }},
         _PROFILE_TAG + " latest"
     )
@@ -162,6 +165,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :mod:`v2023_09_02_preview.models<azure.mgmt.containerservice.v2023_09_02_preview.models>`
            * 2023-10-01: :mod:`v2023_10_01.models<azure.mgmt.containerservice.v2023_10_01.models>`
            * 2023-10-02-preview: :mod:`v2023_10_02_preview.models<azure.mgmt.containerservice.v2023_10_02_preview.models>`
+           * 2023-11-01: :mod:`v2023_11_01.models<azure.mgmt.containerservice.v2023_11_01.models>`
         """
         if api_version == '2017-07-01':
             from .v2017_07_01 import models
@@ -373,6 +377,9 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview import models
             return models
+        elif api_version == '2023-11-01':
+            from .v2023_11_01 import models
+            return models
         raise ValueError("API version {} is not available".format(api_version))
 
     @property
@@ -442,6 +449,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_09_02_preview.operations.AgentPoolsOperations>`
            * 2023-10-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_10_01.operations.AgentPoolsOperations>`
            * 2023-10-02-preview: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_10_02_preview.operations.AgentPoolsOperations>`
+           * 2023-11-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2023_11_01.operations.AgentPoolsOperations>`
         """
         api_version = self._get_api_version('agent_pools')
         if api_version == '2019-02-01':
@@ -570,6 +578,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_10_01.operations import AgentPoolsOperations as OperationClass
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview.operations import AgentPoolsOperations as OperationClass
+        elif api_version == '2023-11-01':
+            from .v2023_11_01.operations import AgentPoolsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'agent_pools'".format(api_version))
         self._config.api_version = api_version
@@ -708,6 +718,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_09_02_preview.operations.MaintenanceConfigurationsOperations>`
            * 2023-10-01: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_10_01.operations.MaintenanceConfigurationsOperations>`
            * 2023-10-02-preview: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_10_02_preview.operations.MaintenanceConfigurationsOperations>`
+           * 2023-11-01: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2023_11_01.operations.MaintenanceConfigurationsOperations>`
         """
         api_version = self._get_api_version('maintenance_configurations')
         if api_version == '2020-12-01':
@@ -808,6 +819,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_10_01.operations import MaintenanceConfigurationsOperations as OperationClass
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview.operations import MaintenanceConfigurationsOperations as OperationClass
+        elif api_version == '2023-11-01':
+            from .v2023_11_01.operations import MaintenanceConfigurationsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'maintenance_configurations'".format(api_version))
         self._config.api_version = api_version
@@ -956,6 +969,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_09_02_preview.operations.ManagedClustersOperations>`
            * 2023-10-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_10_01.operations.ManagedClustersOperations>`
            * 2023-10-02-preview: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_10_02_preview.operations.ManagedClustersOperations>`
+           * 2023-11-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2023_11_01.operations.ManagedClustersOperations>`
         """
         api_version = self._get_api_version('managed_clusters')
         if api_version == '2018-03-31':
@@ -1088,6 +1102,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_10_01.operations import ManagedClustersOperations as OperationClass
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview.operations import ManagedClustersOperations as OperationClass
+        elif api_version == '2023-11-01':
+            from .v2023_11_01.operations import ManagedClustersOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'managed_clusters'".format(api_version))
         self._config.api_version = api_version
@@ -1199,6 +1215,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :class:`Operations<azure.mgmt.containerservice.v2023_09_02_preview.operations.Operations>`
            * 2023-10-01: :class:`Operations<azure.mgmt.containerservice.v2023_10_01.operations.Operations>`
            * 2023-10-02-preview: :class:`Operations<azure.mgmt.containerservice.v2023_10_02_preview.operations.Operations>`
+           * 2023-11-01: :class:`Operations<azure.mgmt.containerservice.v2023_11_01.operations.Operations>`
         """
         api_version = self._get_api_version('operations')
         if api_version == '2018-03-31':
@@ -1331,6 +1348,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_10_01.operations import Operations as OperationClass
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview.operations import Operations as OperationClass
+        elif api_version == '2023-11-01':
+            from .v2023_11_01.operations import Operations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'operations'".format(api_version))
         self._config.api_version = api_version
@@ -1393,6 +1412,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_09_02_preview.operations.PrivateEndpointConnectionsOperations>`
            * 2023-10-01: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_10_01.operations.PrivateEndpointConnectionsOperations>`
            * 2023-10-02-preview: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_10_02_preview.operations.PrivateEndpointConnectionsOperations>`
+           * 2023-11-01: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2023_11_01.operations.PrivateEndpointConnectionsOperations>`
         """
         api_version = self._get_api_version('private_endpoint_connections')
         if api_version == '2020-06-01':
@@ -1501,6 +1521,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_10_01.operations import PrivateEndpointConnectionsOperations as OperationClass
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview.operations import PrivateEndpointConnectionsOperations as OperationClass
+        elif api_version == '2023-11-01':
+            from .v2023_11_01.operations import PrivateEndpointConnectionsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'private_endpoint_connections'".format(api_version))
         self._config.api_version = api_version
@@ -1561,6 +1583,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_09_02_preview.operations.PrivateLinkResourcesOperations>`
            * 2023-10-01: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_10_01.operations.PrivateLinkResourcesOperations>`
            * 2023-10-02-preview: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_10_02_preview.operations.PrivateLinkResourcesOperations>`
+           * 2023-11-01: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2023_11_01.operations.PrivateLinkResourcesOperations>`
         """
         api_version = self._get_api_version('private_link_resources')
         if api_version == '2020-09-01':
@@ -1665,6 +1688,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_10_01.operations import PrivateLinkResourcesOperations as OperationClass
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview.operations import PrivateLinkResourcesOperations as OperationClass
+        elif api_version == '2023-11-01':
+            from .v2023_11_01.operations import PrivateLinkResourcesOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'private_link_resources'".format(api_version))
         self._config.api_version = api_version
@@ -1725,6 +1750,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_09_02_preview.operations.ResolvePrivateLinkServiceIdOperations>`
            * 2023-10-01: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_10_01.operations.ResolvePrivateLinkServiceIdOperations>`
            * 2023-10-02-preview: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_10_02_preview.operations.ResolvePrivateLinkServiceIdOperations>`
+           * 2023-11-01: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2023_11_01.operations.ResolvePrivateLinkServiceIdOperations>`
         """
         api_version = self._get_api_version('resolve_private_link_service_id')
         if api_version == '2020-09-01':
@@ -1829,6 +1855,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_10_01.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
+        elif api_version == '2023-11-01':
+            from .v2023_11_01.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'resolve_private_link_service_id'".format(api_version))
         self._config.api_version = api_version
@@ -1882,6 +1910,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_09_02_preview.operations.SnapshotsOperations>`
            * 2023-10-01: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_10_01.operations.SnapshotsOperations>`
            * 2023-10-02-preview: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_10_02_preview.operations.SnapshotsOperations>`
+           * 2023-11-01: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2023_11_01.operations.SnapshotsOperations>`
         """
         api_version = self._get_api_version('snapshots')
         if api_version == '2021-08-01':
@@ -1972,6 +2001,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_10_01.operations import SnapshotsOperations as OperationClass
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview.operations import SnapshotsOperations as OperationClass
+        elif api_version == '2023-11-01':
+            from .v2023_11_01.operations import SnapshotsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'snapshots'".format(api_version))
         self._config.api_version = api_version
@@ -2002,6 +2033,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2023_09_02_preview.operations.TrustedAccessRoleBindingsOperations>`
            * 2023-10-01: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2023_10_01.operations.TrustedAccessRoleBindingsOperations>`
            * 2023-10-02-preview: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2023_10_02_preview.operations.TrustedAccessRoleBindingsOperations>`
+           * 2023-11-01: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2023_11_01.operations.TrustedAccessRoleBindingsOperations>`
         """
         api_version = self._get_api_version('trusted_access_role_bindings')
         if api_version == '2022-04-02-preview':
@@ -2046,6 +2078,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_10_01.operations import TrustedAccessRoleBindingsOperations as OperationClass
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview.operations import TrustedAccessRoleBindingsOperations as OperationClass
+        elif api_version == '2023-11-01':
+            from .v2023_11_01.operations import TrustedAccessRoleBindingsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'trusted_access_role_bindings'".format(api_version))
         self._config.api_version = api_version
@@ -2076,6 +2110,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2023-09-02-preview: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2023_09_02_preview.operations.TrustedAccessRolesOperations>`
            * 2023-10-01: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2023_10_01.operations.TrustedAccessRolesOperations>`
            * 2023-10-02-preview: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2023_10_02_preview.operations.TrustedAccessRolesOperations>`
+           * 2023-11-01: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2023_11_01.operations.TrustedAccessRolesOperations>`
         """
         api_version = self._get_api_version('trusted_access_roles')
         if api_version == '2022-04-02-preview':
@@ -2120,6 +2155,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
             from .v2023_10_01.operations import TrustedAccessRolesOperations as OperationClass
         elif api_version == '2023-10-02-preview':
             from .v2023_10_02_preview.operations import TrustedAccessRolesOperations as OperationClass
+        elif api_version == '2023-11-01':
+            from .v2023_11_01.operations import TrustedAccessRolesOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'trusted_access_roles'".format(api_version))
         self._config.api_version = api_version