+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/generated_samples/activestamp/azure_iaas_vm/provision_ilr.py
@@ -48,6 +48,6 @@ def main():
     )
 
 
-# x-ms-original-file: specification/recoveryservicesbackup/resource-manager/Microsoft.RecoveryServices/stable/2023-04-01/examples/AzureIaasVm/Provision_Ilr.json
+# x-ms-original-file: specification/recoveryservicesbackup/resource-manager/Microsoft.RecoveryServices/stable/2023-06-01/examples/AzureIaasVm/Provision_Ilr.json
 if __name__ == "__main__":
     main()