+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -1,11 +1,28 @@
 # Release History
 
+## 1.0.0b22 (Unreleased)
+
+### Features Added
+
+- Add live metrics skeleton + swagger definitions
+    ([#33983](https://github.com/Azure/azure-sdk-for-python/pull/33983))
+- Only create temporary folder if local storage is enabled without storage directory.
+    ([#34061](https://github.com/Azure/azure-sdk-for-python/pull/34061))
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 1.0.0b21 (2024-01-16)
 
 ### Other Changes
 
 - Update to OTel SKD/API 1.21
     ([#33864](https://github.com/Azure/azure-sdk-for-python/pull/33864))
+- Update Django sample
+    ([#33834](https://github.com/Azure/azure-sdk-for-python/pull/33834))
 
 ## 1.0.0b20 (2024-01-04)
 
@@ -22,7 +39,7 @@
 - Implement distro detection for statsbeat feature
     ([#33761](https://github.com/Azure/azure-sdk-for-python/pull/33761))
 - Use empty resource for statsbeat `MeterProvider`
-    ([#33761](https://github.com/Azure/azure-sdk-for-python/pull/33761))
+    ([#33768](https://github.com/Azure/azure-sdk-for-python/pull/33768))
 
 ## 1.0.0b19 (2023-11-20)